requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
geographiclib==2.0
geopy==2.4.1
psycopg2==2.9.9
sqlparse==0.4.4
